+++ src/main/java/springboot/shoppingmall/pay/type/kakakopay/web/KakaoPayCancelRequest.java
@@ -1,4 +1,4 @@
-package springboot.shoppingmall.pay.web;
+package springboot.shoppingmall.pay.type.kakakopay.web;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
